+++ common/src/main/java/io/github/jamalam360/reaping/ReapingExpectPlatform.java
@@ -4,7 +4,6 @@ import dev.architectury.injectables.annotations.ExpectPlatform;
 import io.github.jamalam360.reaping.config.ReapingConfig;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.damage.DamageSource;
-import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.loot.context.LootContext;
 
 public class ReapingExpectPlatform {
@@ -19,10 +18,12 @@ public class ReapingExpectPlatform {
     }
 
     @ExpectPlatform
-    public static  LootContext.Builder getLootContextBuilder(LivingEntity entity, boolean causedByPlayer, DamageSource source) {
+    public static LootContext.Builder getLootContextBuilder(LivingEntity entity, boolean causedByPlayer, DamageSource source) {
         throw new AssertionError();
     }
 
     @ExpectPlatform
-    public static void setScale(PlayerEntity player, float scale) { throw new AssertionError(); }
+    public static void setScale(LivingEntity entity, float scale) {
+        throw new AssertionError();
+    }
 }