+++ common/src/main/java/io/github/jamalam360/reaping/ReapingMod.java
@@ -77,7 +77,6 @@ public class ReapingMod {
                                         .statusEffect(new StatusEffectInstance(StatusEffects.NAUSEA), 1)
                                         .build()
                                 )
-                                .rarity(Rarity.UNCOMMON)
                 )
         );
     }